jeez, a very large push. a full reorganization that is prepping for this to be modularized somewhat and just give the whole system a facelift.
deleted all the old states.

we need to do a verification pass that we didnt miss anything systemic in this pass, but I think I got it all.

expect this to be the biggest push in the restructure, but I think a big refactor for the gm and state system is due as well

just got it compiling, have not tested, its late but will hopefully finish the loop tomorrow